Alter ccv type

+++ db/migrate/20180220040608_create_cards.rb
@@ -4,7 +4,7 @@ class CreateCards < ActiveRecord::Migration[5.1]
       t.string :holdername
       t.integer :number
       t.date :expirationDate
-      t.decimal :cvv
+      t.integer :cvv
     end
   end
 end
